feat: 공통 네비게이션 바 구현
- 모든 페이지에서 통일된 네비게이션 바 제공
- 로그인 상태에 따른 조건부 렌더링 (로그인 버튼 ↔ 사용자 아바타)
- 기존 페이지별 개별 헤더 제거로 코드 중복 해결
- 반응형 디자인 및 스크롤 효과 유지

주요 변경사항:
• NavigationBar 컴포넌트 신규 생성
• 루트 레이아웃에 공통 네비게이션 추가
• HomePage, DashboardPage 헤더 영역 제거
• 로그인/회원가입 페이지 네비게이션 정리
• 사용자 인증 상태 기반 UI 조건부 표시
