+++ app/dashboard/DashboardPage.tsx
@@ -3,30 +3,14 @@
 import { useState, useEffect, useCallback } from "react";
 import { useRouter } from "next/navigation";
 import DocumentUpload from "@/components/document-upload";
-import LanguageSelector from "@/components/language-selector";
 import { useLanguage } from "@/contexts/language-context";
-import { signOut as serverSignOut } from "@/app/auth/actions";
 import { deleteDocument } from "@/app/actions/document";
 import { Button } from "@/components/ui/button";
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import {
-  ArrowLeft,
-  LogOut,
-  Settings,
-  User,
   Plus,
   FolderOpen,
 } from "lucide-react";
-import Link from "next/link";
 import { toast } from "sonner";
 import type { User as SupabaseUser } from "@supabase/supabase-js";
 
@@ -63,31 +47,6 @@ export default function DashboardPage({ user }: DashboardPageProps) {
     sortOptions: { field: "created_at", order: "desc" },
     counts: { all: 0, draft: 0, published: 0, completed: 0, expired: 0 },
   });
-
-  const handleSignOut = async () => {
-    // Use server action for sign out - it handles redirect automatically
-    const formData = new FormData();
-    await serverSignOut(formData);
-  };
-
-  const getUserDisplayName = () => {
-    return (
-      user.user_metadata?.name ||
-      user.user_metadata?.full_name ||
-      user.email?.split("@")[0] ||
-      "User"
-    );
-  };
-
-  const getUserInitials = () => {
-    const name = getUserDisplayName();
-    return name
-      .split(" ")
-      .map((n) => n[0])
-      .join("")
-      .toUpperCase()
-      .slice(0, 2);
-  };
 
   // Load documents
   const loadDocuments = useCallback(async () => {
@@ -221,67 +180,10 @@ export default function DashboardPage({ user }: DashboardPageProps) {
   return (
     <div className="container mx-auto px-4 py-8">
       <div className="max-w-6xl mx-auto">
-        {/* Header */}
-        <div className="flex justify-between items-center mb-8">
-          <Link href="/">
-            <Button variant="ghost" size="sm" className="gap-2">
-              <ArrowLeft className="h-4 w-4" />
-              <span>{t("dashboard.backToHome")}</span>
-            </Button>
-          </Link>
-
-          <div className="flex items-center gap-4">
-            <LanguageSelector />
-
-            <DropdownMenu>
-              <DropdownMenuTrigger asChild>
-                <Button
-                  variant="ghost"
-                  className="relative h-10 w-10 rounded-full"
-                >
-                  <Avatar className="h-10 w-10">
-                    <AvatarImage
-                      src={user.user_metadata?.avatar_url}
-                      alt={getUserDisplayName()}
-                    />
-                    <AvatarFallback>{getUserInitials()}</AvatarFallback>
-                  </Avatar>
-                </Button>
-              </DropdownMenuTrigger>
-              <DropdownMenuContent className="w-56" align="end" forceMount>
-                <DropdownMenuLabel className="font-normal">
-                  <div className="flex flex-col space-y-1">
-                    <p className="text-sm font-medium leading-none">
-                      {getUserDisplayName()}
-                    </p>
-                    <p className="text-xs leading-none text-muted-foreground">
-                      {user.email}
-                    </p>
-                  </div>
-                </DropdownMenuLabel>
-                <DropdownMenuSeparator />
-                <DropdownMenuItem>
-                  <User className="mr-2 h-4 w-4" />
-                  <span>{t("dashboard.profile")}</span>
-                </DropdownMenuItem>
-                <DropdownMenuItem>
-                  <Settings className="mr-2 h-4 w-4" />
-                  <span>{t("dashboard.settings")}</span>
-                </DropdownMenuItem>
-                <DropdownMenuSeparator />
-                <DropdownMenuItem onClick={handleSignOut}>
-                  <LogOut className="mr-2 h-4 w-4" />
-                  <span>{t("dashboard.signOut")}</span>
-                </DropdownMenuItem>
-              </DropdownMenuContent>
-            </DropdownMenu>
-          </div>
-        </div>
-
         {/* Welcome Section */}
         <div className="mb-8">
           <h1 className="text-3xl font-bold tracking-tight mb-2">
-            {t("dashboard.welcome", { name: getUserDisplayName() })}
+            {t("dashboard.welcome", { name: user.user_metadata?.name || user.email?.split("@")[0] || "User" })}
           </h1>
           <p className="text-muted-foreground">{t("dashboard.description")}</p>
         </div>
