+++ components/navigation-bar.tsx
@@ -1,0 +1,196 @@
+"use client"
+
+import { useEffect, useState } from "react"
+import Link from "next/link"
+import { FileSignature, User, Settings, LogOut } from "lucide-react"
+import { Button } from "@/components/ui/button"
+import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu"
+import { useLanguage } from "@/contexts/language-context"
+import { ThemeToggle } from "@/components/theme-toggle"
+import { LanguageSelector } from "@/components/language-selector"
+import { getCurrentUserClient } from "@/app/auth/queries"
+import { signOut as serverSignOut } from "@/app/auth/actions"
+import { cn } from "@/lib/utils"
+import type { User as SupabaseUser } from "@supabase/supabase-js"
+
+interface NavigationBarProps {
+  className?: string
+}
+
+export function NavigationBar({ className }: NavigationBarProps) {
+  const { t } = useLanguage()
+  const [user, setUser] = useState<SupabaseUser | null>(null)
+  const [isLoading, setIsLoading] = useState(true)
+  const [scrolled, setScrolled] = useState(false)
+
+  // Get user info on component mount
+  useEffect(() => {
+    const getUser = async () => {
+      try {
+        const currentUser = await getCurrentUserClient()
+        setUser(currentUser)
+      } catch (error) {
+        console.error("Error fetching user:", error)
+        setUser(null)
+      } finally {
+        setIsLoading(false)
+      }
+    }
+
+    getUser()
+  }, [])
+
+  // Handle scroll effect
+  useEffect(() => {
+    const handleScroll = () => {
+      setScrolled(window.scrollY > 50)
+    }
+    window.addEventListener("scroll", handleScroll)
+    return () => window.removeEventListener("scroll", handleScroll)
+  }, [])
+
+  const handleSignOut = async () => {
+    try {
+      const formData = new FormData()
+      await serverSignOut(formData)
+    } catch (error) {
+      console.error("Error signing out:", error)
+    }
+  }
+
+  const getUserDisplayName = () => {
+    if (!user) return "User"
+    return (
+      user.user_metadata?.name ||
+      user.user_metadata?.full_name ||
+      user.email?.split("@")[0] ||
+      "User"
+    )
+  }
+
+  const getUserInitials = () => {
+    const name = getUserDisplayName()
+    return name
+      .split(" ")
+      .map((n) => n[0])
+      .join("")
+      .toUpperCase()
+      .slice(0, 2)
+  }
+
+  // Show loading skeleton during initial load
+  if (isLoading) {
+    return (
+      <header
+        className={cn(
+          "sticky top-0 z-50 w-full transition-all duration-200",
+          "bg-background/80 backdrop-blur-md border-b",
+          className
+        )}
+      >
+        <div className="container mx-auto px-4 py-4">
+          <div className="flex justify-between items-center">
+            <Link href="/" className="flex items-center gap-2">
+              <FileSignature className="h-8 w-8 text-primary" />
+              <span className="font-bold text-xl">{t("app.title")}</span>
+            </Link>
+            <div className="flex items-center gap-4">
+              <ThemeToggle />
+              <LanguageSelector />
+              <div className="w-10 h-10 bg-muted rounded-full animate-pulse" />
+            </div>
+          </div>
+        </div>
+      </header>
+    )
+  }
+
+  return (
+    <header
+      className={cn(
+        "sticky top-0 z-50 w-full transition-all duration-200",
+        scrolled ? "bg-background/80 backdrop-blur-md border-b" : "bg-transparent",
+        className
+      )}
+    >
+      <div className="container mx-auto px-4 py-4">
+        <div className="flex justify-between items-center">
+          {/* Logo */}
+          <Link href="/" className="flex items-center gap-2">
+            <FileSignature className="h-8 w-8 text-primary" />
+            <span className="font-bold text-xl">{t("app.title")}</span>
+          </Link>
+
+          {/* Right side navigation */}
+          <div className="flex items-center gap-4">
+            <ThemeToggle />
+            <LanguageSelector />
+            
+            {user ? (
+              /* Logged in - Show user avatar and dropdown */
+              <DropdownMenu>
+                <DropdownMenuTrigger asChild>
+                  <Button
+                    variant="ghost"
+                    className="relative h-10 w-10 rounded-full"
+                  >
+                    <Avatar className="h-10 w-10">
+                      <AvatarImage
+                        src={user.user_metadata?.avatar_url}
+                        alt={getUserDisplayName()}
+                      />
+                      <AvatarFallback>{getUserInitials()}</AvatarFallback>
+                    </Avatar>
+                  </Button>
+                </DropdownMenuTrigger>
+                <DropdownMenuContent className="w-56" align="end" forceMount>
+                  <DropdownMenuLabel className="font-normal">
+                    <div className="flex flex-col space-y-1">
+                      <p className="text-sm font-medium leading-none">
+                        {getUserDisplayName()}
+                      </p>
+                      <p className="text-xs leading-none text-muted-foreground">
+                        {user.email}
+                      </p>
+                    </div>
+                  </DropdownMenuLabel>
+                  <DropdownMenuSeparator />
+                  <DropdownMenuItem asChild>
+                    <Link href="/dashboard">
+                      <User className="mr-2 h-4 w-4" />
+                      <span>{t("dashboard.profile")}</span>
+                    </Link>
+                  </DropdownMenuItem>
+                  <DropdownMenuItem>
+                    <Settings className="mr-2 h-4 w-4" />
+                    <span>{t("dashboard.settings")}</span>
+                  </DropdownMenuItem>
+                  <DropdownMenuSeparator />
+                  <DropdownMenuItem onClick={handleSignOut}>
+                    <LogOut className="mr-2 h-4 w-4" />
+                    <span>{t("dashboard.signOut")}</span>
+                  </DropdownMenuItem>
+                </DropdownMenuContent>
+              </DropdownMenu>
+            ) : (
+              /* Not logged in - Show login button */
+              <Link href="/login">
+                <Button className="bg-primary hover:bg-primary/90">
+                  {t("login.logIn")}
+                </Button>
+              </Link>
+            )}
+          </div>
+        </div>
+      </div>
+    </header>
+  )
+}
