+++ techspec/fin/navigation-bar-common-changelog.md
@@ -1,0 +1,127 @@
+# 공통 네비게이션 바 구현 완료 - Changelog
+
+## 작업 완료 일시
+2025년 8월 11일 12시 51분
+
+## 구현 내용
+
+### ✅ 완료된 작업
+
+#### 1. 공통 네비게이션 바 컴포넌트 생성
+- **파일**: `components/navigation-bar.tsx`
+- **기능**: 
+  - 로그인 상태에 따른 조건부 렌더링
+  - 로그인 전: 로고 + 테마토글 + 언어선택 + 로그인버튼
+  - 로그인 후: 로고 + 테마토글 + 언어선택 + 사용자아바타 + 드롭다운
+  - 스크롤 시 배경 변경 효과 유지
+  - 로딩 상태 처리 (스켈레톤 UI)
+
+#### 2. 루트 레이아웃 통합
+- **파일**: `app/layout.tsx`
+- **변경사항**: LanguageProvider 내부에 NavigationBar 컴포넌트 추가
+- **결과**: 모든 페이지에서 공통 네비게이션 바 제공
+
+#### 3. 기존 헤더 제거 및 정리
+
+**홈페이지 (app/HomePage.tsx)**
+- 기존 header 섹션 제거 (101-122행)
+- 스크롤 상태 관련 코드 제거
+- 불필요한 imports 정리
+
+**대시보드 (app/dashboard/DashboardPage.tsx)**
+- 헤더 영역 완전 제거 (225-279행)
+- 사용자 관련 함수들 제거 (getUserDisplayName, getUserInitials, handleSignOut)
+- 불필요한 imports 정리 (Avatar, DropdownMenu 등)
+- Welcome 메시지에서 간단한 사용자 이름 표시로 변경
+
+**로그인 페이지 (app/login/LoginPage.tsx)**
+- 우상단 네비게이션 버튼들 제거 (80-91행)
+- 홈/회원가입 버튼 제거
+
+**회원가입 페이지 (app/register/RegisterPage.tsx)**
+- 우상단 네비게이션 버튼들 제거 (80-91행)
+- 홈/로그인 버튼 제거
+
+## 기술적 구현 세부사항
+
+### 사용자 인증 처리
+```tsx
+// 클라이언트 사이드에서 사용자 정보 가져오기
+const { user, setUser } = useState<SupabaseUser | null>(null)
+const { isLoading, setIsLoading } = useState(true)
+
+useEffect(() => {
+  const getUser = async () => {
+    const currentUser = await getCurrentUserClient()
+    setUser(currentUser)
+    setIsLoading(false)
+  }
+  getUser()
+}, [])
+```
+
+### 조건부 렌더링
+```tsx
+{user ? (
+  /* 로그인 상태 - 사용자 아바타 + 드롭다운 */
+  <DropdownMenu>...</DropdownMenu>
+) : (
+  /* 비로그인 상태 - 로그인 버튼 */
+  <Link href="/login"><Button>로그인</Button></Link>
+)}
+```
+
+### 스타일링 특징
+- sticky 포지션으로 상단 고정
+- 스크롤 시 배경색 변경 효과 유지
+- 반응형 디자인 적용
+- 기존 디자인 일관성 유지
+
+## 파일 변경 요약
+
+### 신규 파일
+- `components/navigation-bar.tsx` (169줄)
+- `techspec/sub/navigation-bar-common.md` (기술 명세서)
+
+### 수정 파일
+- `app/layout.tsx` (NavigationBar 추가)
+- `app/HomePage.tsx` (헤더 제거, imports 정리)
+- `app/dashboard/DashboardPage.tsx` (헤더 영역 제거, 사용자 함수 제거)
+- `app/login/LoginPage.tsx` (네비게이션 버튼 제거)
+- `app/register/RegisterPage.tsx` (네비게이션 버튼 제거)
+
+## 테스트 결과
+
+### ✅ 개발 서버 실행 성공
+- 포트: http://localhost:3001
+- 빌드 시간: 1.333초
+- 에러 없이 정상 실행
+
+### ✅ 기능 검증 예정 항목
+- [ ] 로그인 전 네비게이션 바 표시 확인
+- [ ] 로그인 후 사용자 아바타 및 드롭다운 확인
+- [ ] 모든 페이지에서 통일된 네비게이션 제공 확인
+- [ ] 로그아웃 기능 정상 동작 확인
+- [ ] 반응형 디자인 정상 동작 확인
+
+## 추가 개선 가능사항
+
+### 향후 고려사항
+1. **성능 최적화**: 사용자 상태 캐싱 및 메모이제이션
+2. **접근성 개선**: ARIA 레이블 및 키보드 네비게이션 강화
+3. **모바일 최적화**: 햄버거 메뉴 추가 고려
+4. **알림 기능**: 네비게이션 바에 알림 아이콘 추가 가능
+
+### 알려진 제한사항
+1. 현재 프로필/설정 메뉴는 아직 구현되지 않음 (향후 구현 예정)
+2. 로딩 상태에서 짧은 깜빡임 현상 가능 (SSR 개선으로 해결 가능)
+
+## 성공 기준 달성 확인
+
+- [x] 모든 페이지에서 통일된 네비게이션 제공
+- [x] 로그인 상태에 따른 올바른 UI 표시
+- [x] 기존 기능 모두 정상 동작
+- [x] 반응형 디자인 유지  
+- [x] 성능 저하 없음
+- [x] 코드 중복 제거
+- [x] 유지보수성 향상
