+++ techspec/sub/navigation-bar-common.md
@@ -1,0 +1,120 @@
+# 공통 네비게이션 바 구현 기술 명세서
+
+## 개요
+각 페이지마다 개별적으로 구현된 헤더/네비게이션을 공통 컴포넌트로 통합하여 일관된 사용자 경험 제공
+
+## 현재 상황 분석
+
+### 기존 구현 현황
+1. **홈페이지 (app/HomePage.tsx)**
+   - 자체 header 컴포넌트 (101-122행)
+   - 로그인 버튼 포함
+   - 스크롤 시 배경색 변경 효과
+
+2. **대시보드 (app/dashboard/DashboardPage.tsx)**  
+   - 별도 헤더 영역 (221-278행)
+   - 사용자 아바타 + 드롭다운 메뉴
+   - 언어 선택기 + 로그아웃 기능
+
+3. **로그인/회원가입 페이지**
+   - 각각 우상단에 별도 네비게이션 버튼들
+   - 홈으로 돌아가기, 로그인/회원가입 전환 버튼
+
+### 문제점
+- 코드 중복
+- 일관성 없는 디자인
+- 유지보수 어려움
+- 사용자 상태에 따른 네비게이션 로직 분산
+
+## 설계 방안
+
+### 1. 공통 네비게이션 컴포넌트 구조
+```tsx
+interface NavigationBarProps {
+  // 추가 props가 필요한 경우를 대비
+}
+
+// 로그인 상태에 따른 조건부 렌더링
+// - 로그인 전: 로고 + 테마토글 + 언어선택 + 로그인버튼  
+// - 로그인 후: 로고 + 테마토글 + 언어선택 + 사용자아바타
+```
+
+### 2. 사용자 상태 관리
+- `useUser` 훅 활용하여 현재 사용자 정보 획득
+- 클라이언트 사이드 렌더링 고려
+- 로딩 상태 처리
+
+### 3. 배치 위치
+- `app/layout.tsx`의 Provider 내부에 배치
+- 모든 페이지에서 공통으로 표시
+- sticky 포지션으로 상단 고정
+
+## 구현 단계
+
+### Phase 1: 공통 컴포넌트 생성
+- [ ] `components/navigation-bar.tsx` 생성
+- [ ] 기존 대시보드의 사용자 메뉴 로직 이전
+- [ ] 로그인/로그아웃 상태별 조건부 렌더링
+
+### Phase 2: 레이아웃 통합
+- [ ] `app/layout.tsx`에 네비게이션 바 추가
+- [ ] 기존 페이지별 헤더 제거
+
+### Phase 3: 페이지별 정리
+- [ ] HomePage.tsx 헤더 제거
+- [ ] DashboardPage.tsx 헤더 영역 제거  
+- [ ] 로그인/회원가입 페이지 정리
+
+### Phase 4: 테스트 및 검증
+- [ ] 로그인/로그아웃 상태 전환 테스트
+- [ ] 반응형 디자인 확인
+- [ ] 모든 페이지에서 정상 동작 확인
+
+## 기술적 고려사항
+
+### 1. 사용자 인증 상태 처리
+```tsx
+// 클라이언트 컴포넌트에서 사용자 정보 가져오기
+const { user, isLoading } = useUser()
+
+// 로딩 상태 처리
+if (isLoading) return <NavigationSkeleton />
+```
+
+### 2. 스타일링 일관성
+- 기존 HomePage의 header 스타일 기반으로 통합
+- 스크롤 시 배경 변경 효과 유지
+- 대시보드의 사용자 메뉴 스타일 적용
+
+### 3. 접근성 고려
+- 키보드 네비게이션 지원
+- 스크린 리더 호환성
+- ARIA 레이블 적용
+
+## 예상 파일 변경사항
+
+### 신규 파일
+- `components/navigation-bar.tsx`
+
+### 수정 파일
+- `app/layout.tsx` (네비게이션 바 추가)
+- `app/HomePage.tsx` (헤더 제거)  
+- `app/dashboard/DashboardPage.tsx` (헤더 영역 제거)
+- `app/login/LoginPage.tsx` (네비게이션 정리)
+- `app/register/RegisterPage.tsx` (네비게이션 정리)
+
+## 마이그레이션 계획
+
+1. 기존 코드 백업 (git branch 생성)
+2. 단계별 구현 및 테스트
+3. 각 페이지별 점진적 적용
+4. 최종 통합 테스트
+5. 코드 리뷰 및 배포
+
+## 성공 기준
+
+- [ ] 모든 페이지에서 통일된 네비게이션 제공
+- [ ] 로그인 상태에 따른 올바른 UI 표시
+- [ ] 기존 기능 모두 정상 동작
+- [ ] 반응형 디자인 유지
+- [ ] 성능 저하 없음
